add github service module
