+++ config/application.rb
@@ -10,6 +10,7 @@ Bundler.require(*Rails.groups)
 
 module AwhiMai
   class Application < Rails::Application
+    config.autoload_paths += %W[#{config.root}/lib]
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 5.2
     config.generators.javascript_engine = :js
