+++ config/initializers/custom_modules.rb
@@ -1,0 +1,1 @@
+require 'github_service'
